requirements.txt
# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
Pillow>=9.0.0
# Enhanced file format support
pandas>=1.5.0  # For CSV/Excel processing
openpyxl>=3.1.0  # For Excel files
markdown>=3.4.0  # For markdown files
beautifulsoup4>=4.12.0  # For HTML parsing
lxml>=4.9.0  # For XML processing
odfpy>=1.4.0  # For ODT files
chardet>=5.0.0  # For text encoding detection

# OCR and Computer Vision
pytesseract>=0.3.9
opencv-python>=4.7.0
easyocr>=1.6.0

# NLP and AI Models (Free)
transformers>=4.30.0
torch>=1.13.0
sentence-transformers>=2.2.0
# Translation and language detection
sacrebleu>=2.3.0  # For translation evaluation
langdetect>=1.0.9  # For better language detection

# Vector Storage (FAISS - now optional)
faiss-cpu>=1.7.0

# MongoDB Database
motor>=3.3.0
pymongo>=4.6.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0
python-multipart>=0.0.5
jinja2>=3.1.0

# Utilities
numpy>=1.21.0
pandas>=1.5.0
pydantic>=2.0.0
aiofiles>=23.0.0
scikit-learn>=1.0.0
requests>=2.28.0

# Additional utilities for Windows
python-magic-bin>=0.4.14; sys_platform == "win32"

# HTTP clients
httpx>=0.24.0
aiohttp>=3.8.0

# Development tools (optional)
tqdm>=4.64.0
coloredlogs>=15.0
python-dotenv>=1.0.0
